fixed parsing amount string

+++ src/utils/utils.ts
@@ -1,7 +1,7 @@
 import { Bank } from "../types/bank.type";
 
 export const parseLocaleString = (num: number, sign: boolean | never) => {
-  const parsedNumber = parseInt(num.toFixed(2)).toLocaleString();
+  const parsedNumber = parseFloat(num.toFixed(2)).toLocaleString();
   const finalReturn = `${sign && "₱"}${parsedNumber}`;
   return finalReturn;
 }
